requirements.txt
asgiref==3.3.1
attrs==20.3.0
certifi==2020.12.5
chardet==4.0.0
Django==3.1.6
django-environ==0.4.5
django-filter==2.4.0
django-simple-history==2.12.0
djangorestframework==3.12.2
djangorestframework-csv==2.1.0
drf-spectacular==0.13.2
flake8==3.8.4
idna==2.10
inflection==0.5.1
isort==5.7.0
jsonschema==3.2.0
mccabe==0.6.1
pycodestyle==2.6.0
pyflakes==2.2.0
pyrsistent==0.17.3
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
unicodecsv==0.14.1
uritemplate==3.0.1
urllib3==1.26.3
